(#3)route 분리- 2번쨰
